requirements.txt
asgiref
Django
django-cors-headers
django-environ
django-filter
djangorestframework
Markdown
pillow
pip
psycopg2
psycopg2-binary
python-dotenv
sqlparse
whitenoise